# src/main/apigee/apiproxies/TradingAPI-v1/apiproxy/resources/oas/apispec.yaml
openapi: 3.0.0
info:
  title: Energy Trading Management API
  description: >-
    API for energy trading management, with methods to execute, monitor and
    track trades.
  version: 1.0.0
  contact:
    name: Train Support
    url: https://example.com/support
paths:
  /trades:
    get:
      description: List all trades.
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
          description: Bad Request
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
          description: Internal Server Error
      security:
        - OAuth2:
            - write
      tags:
        - trades
    post:
      description: Creates a new trade and returns the created trade.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Trade'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
          description: OK
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
          description: Bad Request
      security:
        - OAuth2:
            - write
      summary: Create a new trade
      tags:
        - trades
  /trades/{tradeId}:
    delete:
      description: Deletes a specific trade by its unique identifier.
      operationId: deleteTrade
      responses:
        '200':
          description: The trade was deleted successfully.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
          description: The trade with the specified identifier does not exist.
      security:
        - OAuth2:
            - write
      summary: Delete a specific trade.
      tags:
        - trades
    get:
      description: Get the details of a specific trade.
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
          description: The details of the trade.
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
          description: The trade was not found.
      security:
        - OAuth2:
            - write
      tags:
        - trades
    parameters:
      - description: The unique identifier of the trade.
        in: path
        name: tradeId
        required: true
        schema:
          type: string
    put:
      description: Update the details of a specific trade.
      operationId: updateTrade
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Trade'
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Trade'
          description: The updated trade.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
          description: Bad Request
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
          description: Not Found
      security:
        - OAuth2:
            - write
      tags:
        - trades
components:
  schemas:
    Country:
      description: The country in which the trade is taking place.
      properties:
        name:
          description: The name of the country.
          type: string
      type: object
    EnergyProvider:
      description: The energy provider involved in the trade.
      properties:
        name:
          description: The name of the energy provider.
          type: string
      required:
        - name
      type: object
    Trade:
      description: Trade object represents a single energy trade.
      properties:
        country:
          $ref: '#/components/schemas/Country'
          description: The country in which the trade is taking place.
        energyProvider:
          $ref: '#/components/schemas/EnergyProvider'
          description: The energy provider involved in the trade.
        price:
          description: The price of the trade.
          type: number
        quantity:
          description: The quantity of energy traded.
          type: number
        trackingInformation:
          description: Tracking information for the trade.
          type: string
        tradeId:
          description: The unique identifier of the trade.
          type: string
      type: object
    error:
      description: Error message.
      properties:
        errorMessage:
          description: Error message.
          type: string
      type: object
  securitySchemes:
    OAuth2:
      type: oauth2
      description: OAuth 2.0 authorization code following RFC8725 best practices.
      flows:
        authorizationCode:
          authorizationUrl: https://example.com/oauth/authorize
          tokenUrl: https://example.com/oauth/token
          scopes:
            read: Read access
            write: Write access